refactor: change badge package and simplify printResult
