+++ kotlin-christmas/src/main/kotlin/christmas/enums/Badge.kt
@@ -1,4 +1,4 @@
-package christmas.data
+package christmas.enums
 
 enum class Badge(
     val badgeName: String,
